Use `bundler` module resolution

+++ fixtures/react.tsx
@@ -1,5 +1,7 @@
 import {useState} from 'react'
 
+import {double} from './utils'
+
 interface FooProps {
   num: number
   str: string
@@ -11,7 +13,7 @@ export default function Foo(props: FooProps): JSX.Element {
   return (
     <div>
       {count}
-      {num}
+      {double(num)}
       {str}
     </div>
   )
